// packages/destination-actions/src/destinations/emarsys/addToContactList/generated-types.ts
// Generated file. DO NOT MODIFY IT BY HAND.

export interface Payload {
  /**
   * The Id of the contactlist
   */
  contactlistid: number
  /**
   * The field to use to find the contact
   */
  key_field: string
  /**
   * Value for the key field used to find the contact. E.g. the email address
   */
  key_value: string
}